Finished HTML and CSS/comments for HTML
Created the form and added in its components and comments to go along with them. Some of the components include the input fields and their labels which work to create a space for user input, the datalist for the title lable which offers some suggestions as to their title, and etc. I also finished the CSS, the styles were just there to inform the user whether their information that was input was valid or invalid.